Make inspect() async and self-contained with full pipeline
Move schema generation, contract hash computation, and workflow finalization
into inspect() so it returns a fully finalized state in one call. This fixes
a bug where finalizeWorkflows() ran before contract hashes existed, causing
workflow step hashes to use empty strings.

Consolidate all contract/version logic (types, validation, manifest helpers)
from CLI into @pikku/inspector. Clean up inspector's public API by removing
internal-only exports. Make graph hash computation explicit by picking
structural fields instead of excluding metadata fields.
