+++ packages/cli/src/functions/commands/versions-check.ts
@@ -3,7 +3,7 @@ import { pikkuSessionlessFunc } from '#pikku'
 import { ErrorCode } from '@pikku/inspector'
 import {
   loadManifest,
-  buildCurrentContracts,
+  extractContractsFromMeta,
   validateContracts,
 } from '../../utils/contract-versions.js'
 
@@ -13,50 +13,19 @@ export const pikkuVersionsCheck = pikkuSessionlessFunc<void, void>({
 
     const manifest = await loadManifest(manifestPath)
     if (!manifest) {
-      logger.error(
-        `[${ErrorCode.MANIFEST_MISSING}] Version manifest not found at ${manifestPath}. Run 'pikku versions init' to create one.`
+      throw new Error(
+        `[${ErrorCode.MANIFEST_MISSING}] Version manifest not found at ${manifestPath}. Run 'pikku init' to create one.`
       )
-      process.exit(1)
     }
 
     const visitState = await getInspectorState()
-
-    const { generateSchemas, generateZodSchemas } = await import(
-      '../../utils/schema-generator.js'
-    )
-
-    const zodSchemas = await generateZodSchemas(
-      logger,
-      visitState.schemaLookup,
-      visitState.functions.typesMap
-    )
-
-    const schemas = await generateSchemas(
-      logger,
-      config.tsconfig,
-      visitState.functions.typesMap,
-      visitState.functions.meta,
-      visitState.http.meta,
-      config.schemasFromTypes,
-      config.schema?.additionalProperties,
-      visitState.schemaLookup
-    )
-
-    const allSchemas = { ...schemas, ...zodSchemas }
-
-    const contracts = buildCurrentContracts(
-      visitState.functions.meta,
-      allSchemas,
-      visitState.functions.typesMap
-    )
+    const contracts = extractContractsFromMeta(visitState.functions.meta)
 
     const result = validateContracts(manifest, contracts)
 
     if (!result.valid) {
-      for (const error of result.errors) {
-        logger.error(`[${error.code}] ${error.message}`)
-      }
-      process.exit(1)
+      const messages = result.errors.map((e) => `[${e.code}] ${e.message}`)
+      throw new Error(messages.join('\n'))
     }
 
     logger.info('Version manifest check passed.')
