+++ packages/cli/src/functions/wirings/ai-agent/serialize-ai-agent-types.ts
@@ -21,6 +21,8 @@ export const pikkuAIAgent = (
   return agent
 }
 
+//TODO: SWitch RequiredServices to be the second generic and default it to Services, and Event to be the first generic and default it to unknown, to match the order of generics in PikkuMiddlewareHooks. 
+// should also pass in state as a type here
 export const pikkuAIMiddleware = <
   RequiredServices extends Services = Services,
   Event = unknown,
