+++ packages/cli/src/functions/wirings/workflow/pikku-command-workflow.ts
@@ -1,6 +1,5 @@
 import { pikkuSessionlessFunc } from '#pikku'
-import { convertDslToGraph, ErrorCode } from '@pikku/inspector'
-import type { WorkflowsMeta } from '@pikku/core/workflow'
+import { ErrorCode } from '@pikku/inspector'
 import { writeFileInDir } from '../../../utils/file-writer.js'
 import { logCommandInfoAndTime } from '../../../middleware/log-command-info-and-time.js'
 import { serializeWorkflowTypes } from './serialize-workflow-types.js'
@@ -13,11 +12,6 @@ import {
   stripVerboseFields,
   hasVerboseFields,
 } from '../../../utils/strip-verbose-meta.js'
-import {
-  computeGraphHash,
-  computeStepHashes,
-} from '../../../utils/workflow-hash.js'
-import { stampVersionsOnGraph } from '../../../utils/stamp-versions-on-graph.js'
 import { join } from 'path'
 
 export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
@@ -37,16 +31,9 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
     const { workflows, functions: functionState } = visitState
     const { typesMap } = functionState
 
-    // Get all workflow names (both DSL and graph-based)
-    const dslWorkflowNames = Object.keys(workflows.meta)
-    const graphWorkflowNames = Object.keys(workflows.graphMeta)
-    const allWorkflowNames = [
-      ...new Set([...dslWorkflowNames, ...graphWorkflowNames]),
-    ]
-
+    const allWorkflowNames = Object.keys(workflows.graphMeta)
     const hasWorkflows = allWorkflowNames.length > 0
 
-    // Validate that workflowService is configured if workflows are defined
     if (hasWorkflows) {
       const hasWorkflowState =
         visitState.serviceAggregation.allSingletonServices.includes(
@@ -63,45 +50,8 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       }
     }
 
-    // Generate individual JSON files for each workflow (convert DSL to graph format)
     if (hasWorkflows && workflowMetaDir) {
-      // Write individual JSON files for DSL workflows
-      const dslMeta = workflows.meta as WorkflowsMeta
-      for (const [name, meta] of Object.entries(dslMeta)) {
-        const graphMeta = convertDslToGraph(name, meta)
-
-        stampVersionsOnGraph(graphMeta, functionState.meta)
-        computeStepHashes(graphMeta, functionState.meta)
-        graphMeta.graphHash = computeGraphHash(graphMeta)
-
-        // Write minimal version (runtime-only fields)
-        const minimalMeta = stripVerboseFields(graphMeta)
-        const minimalPath = join(workflowMetaDir, `${name}.gen.json`)
-        await writeFileInDir(
-          logger,
-          minimalPath,
-          JSON.stringify(minimalMeta, null, 2),
-          { ignoreModifyComment: true }
-        )
-
-        // Write verbose version only if it has additional fields
-        if (hasVerboseFields(graphMeta)) {
-          const verbosePath = join(workflowMetaDir, `${name}-verbose.gen.json`)
-          await writeFileInDir(
-            logger,
-            verbosePath,
-            JSON.stringify(graphMeta, null, 2),
-            { ignoreModifyComment: true }
-          )
-        }
-      }
-
       for (const [name, graphMeta] of Object.entries(workflows.graphMeta)) {
-        stampVersionsOnGraph(graphMeta, functionState.meta)
-        computeStepHashes(graphMeta, functionState.meta)
-        graphMeta.graphHash = computeGraphHash(graphMeta)
-
-        // Write minimal version (runtime-only fields)
         const minimalMeta = stripVerboseFields(graphMeta)
         const minimalPath = join(workflowMetaDir, `${name}.gen.json`)
         await writeFileInDir(
@@ -111,7 +61,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
           { ignoreModifyComment: true }
         )
 
-        // Write verbose version only if it has additional fields
         if (hasVerboseFields(graphMeta)) {
           const verbosePath = join(workflowMetaDir, `${name}-verbose.gen.json`)
           await writeFileInDir(
@@ -124,7 +73,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       }
     }
 
-    // Generate workflow meta aggregation file
     if (workflowsWiringMetaFile && workflowMetaDir) {
       await writeFileInDir(
         logger,
@@ -140,7 +88,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       )
     }
 
-    // Generate workflow registration (meta + DSL workflow registrations)
     const metaImportPath = getFileImportRelativePath(
       workflowsWiringFile,
       workflowsWiringMetaFile,
@@ -161,7 +108,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       )
     )
 
-    // Generate workflow types (DSL + graph helpers in one file)
     const functionTypesImportPath = getFileImportRelativePath(
       workflowTypesFile,
       functionTypesFile,
@@ -188,8 +134,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       )
     )
 
-    // Generate workflow map (I/O types for type-safe client)
-    // Always generate even if empty - RPC map imports WorkflowMap
     await writeFileInDir(
       logger,
       workflowMapDeclarationFile,
@@ -204,7 +148,6 @@ export const pikkuWorkflow = pikkuSessionlessFunc<void, boolean | undefined>({
       )
     )
 
-    // Generate workflow workers if configured
     if (config.workflows) {
       if (config.workflows.singleQueue) {
         const workflowPath = join(config.rootDir, config.workflows.path)
