+++ packages/cli/src/utils/serialize-schemas.ts
