+++ packages/inspector/src/utils/hash.ts
