+++ packages/inspector/src/utils/workflow/graph/index.ts
@@ -4,3 +4,4 @@
 export * from './workflow-graph.types.js'
 export { serializeWorkflowGraph } from './serialize-workflow-graph.js'
 export { convertDslToGraph } from './convert-dsl-to-graph.js'
+export { finalizeWorkflows } from './finalize-workflows.js'
